Organize cli-config and replace previous override implementation
This allows for defining custom CLI options and hooks
