+++ lib/commands/build.js
@@ -5,7 +5,7 @@ const { applyOptions, stripesConfigOptions, buildOptions } = require('./common-o
 const { processError, processStats, cliResolve, cliAliases, emitLintWarnings } = require('../webpack-common');
 
 
-function buildCommand(argv) {
+function buildCommand(argv, plugin) {
   if (argv.dev) {
     console.log('Development config not yet implemented with build');
     return;
@@ -39,7 +39,9 @@ function buildCommand(argv) {
   if (argv.lint) {
     webpackOverrides.push(emitLintWarnings);
   }
-  webpackOverrides.push(platform.getOverride(argv));
+  if (plugin.beforeBuild) {
+    webpackOverrides.push(plugin.beforeBuild);
+  }
 
   // TODO: Check stripes-core location and aliases to warn if build is unsuitable for production.
