+++ lib/commands/serve.js
@@ -4,7 +4,7 @@ const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, serverOptions, stripesConfigOptions, buildOptions } = require('./common-options');
 const { cliResolve, cliAliases, emitLintWarnings } = require('../webpack-common');
 
-function serveCommand(argv) {
+function serveCommand(argv, plugin) {
   if (argv.prod) {
     console.log('Production config not yet implemented with serve');
     return;
@@ -25,10 +25,8 @@ function serveCommand(argv) {
   if (argv.lint) {
     webpackOverrides.push(emitLintWarnings);
   }
-
-  // TODO: Standardize the beforeBuild hook and apply to all build related commands
-  if (argv.beforeBuild && typeof argv.beforeBuild === 'function') {
-    webpackOverrides.push(argv.beforeBuild);
+  if (plugin.beforeBuild) {
+    webpackOverrides.push(plugin.beforeBuild);
   }
 
   console.log('Waiting for webpack to build...');
