+++ lib/platform/stripes-platform.js
@@ -10,7 +10,6 @@ module.exports = class StripesPlatform {
     if (stripesConfigFile) {
       const stripesConfig = require(path.resolve(stripesConfigFile)); // eslint-disable-line
       this.config = mergeConfig(emptyConfig, stripesConfig);
-      this.override = stripesConfig.override;
     } else {
       this.config = defaultConfig;
     }
@@ -60,20 +59,10 @@ module.exports = class StripesPlatform {
   getStripesConfig() {
     const config = Object.assign({}, this.config);
     delete config.aliases;
-    delete config.override; // TODO: Consolidate CLI-specific properties under 'cli' to simplify removal
     return config;
   }
 
   getAliases() {
     return this.config.aliases;
-  }
-
-  getOverride(argv) {
-    // TODO: Establish a suitable interface for applying one or more overrides
-    if (this.override && typeof this.override === 'function') {
-      return this.override(argv);
-    } else {
-      return config => config;
-    }
   }
 };
